refactor(shipping): Phase 10 enforce layout boundaries
- Port LayoutAnalyzer detection logic from Inventory (using-directive +
  symbol-ref handlers; namespace-prefix rules for Domain/Features/
  Infrastructure). Diagnostic IDs SHPLAY001-003 unchanged; severity
  already Error in the descriptor.
- Unskip the 4 NetArchTest LayoutTests + 3 LayoutAnalyzerTests.
- Pre-existing leak surfaced by the Contracts rule: ShipmentStatusChangedEvent
  referenced Shipping.Service.Domain.ShipmentStatus. Switched FromStatus/
  ToStatus to int?/int (wire format already serializes the enum as int
  per existing endpoint tests asserting "ToStatus":{(int)status}); added
  casts at 10 construction sites and rebased 2 CarrierStatusApplierTests
  assertions. Domain layer remains the canonical owner of ShipmentStatus.

+++ shipping-microservice/Shipping.Service/Features/CreateShipmentCommand/CreateShipmentCommandHandler.cs
@@ -69,7 +69,7 @@ internal sealed class CreateShipmentCommandHandler : IEventHandler<ECommerce.Sha
                     shipment.Id,
                     shipment.OrderId,
                     FromStatus: null,
-                    ToStatus: Domain.ShipmentStatus.Pending,
+                    ToStatus: (int)Domain.ShipmentStatus.Pending,
                     OccurredAt: shipment.CreatedAt)
                 {
                     CorrelationId = command.CorrelationId,
